Store fread result in variable with correct type - size_t (cherry picked from commit ee49ae3abc5eb3c7212030aeb6c8732def4878cf)

+++ src/common/mipmap_cache.c
@@ -318,7 +318,7 @@ write_error:
 static int
 dt_mipmap_cache_deserialize(dt_mipmap_cache_t *cache)
 {
-  int32_t rd = 0;
+  size_t rd = 0;
   const dt_mipmap_size_t mip = DT_MIPMAP_2;
   uint32_t *blob = NULL;
   FILE *f = NULL;
